// src/kryptutil-api-out/model/signAlg.ts
/**
 * Kryptutil API
 * JWT, crypt, hash
 *
 * The version of the OpenAPI document: v0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SignAlg { 
    name: string;
    minLen: number;
    description: string;
    usage: string;
}