// src/kryptutil-api-out/model/tokenRequest.ts
/**
 * Kryptutil API
 * JWT, crypt, hash
 *
 * The version of the OpenAPI document: v0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RequestedClaim } from './requestedClaim';


export interface TokenRequest { 
    subject?: string;
    tokenName?: string;
    audience?: string;
    notBefore?: string;
    expiration?: string;
    issuer?: string;
    key?: string;
    alg?: string;
    claims?: Array<RequestedClaim>;
}