// src/kryptutil-api-out/model/requestedClaim.ts
/**
 * Kryptutil API
 * JWT, crypt, hash
 *
 * The version of the OpenAPI document: v0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RequestedClaim { 
    name?: string;
    value?: string;
}